Massive commit and even bigger comment, so sit down first.
Level: Now functions only to read in files and create the map. All objects (Terrain and Units) are stored in the array. Changed a few variable names so they'd be shorter. Also, it assigns units textures when they're read in from the file, rather than giving them an index. This way, GridOccupants Img attribute is actually *used* (and when we draw we just use that stored attribute) and GridOccupants aren't storing information that's only used by other classes.

Main inheritance tree: Added constructors that accepted only x & y values

GameVariables: Stored textures in array (terrain, player units, and enemies each get their own) to cut down on the number of attributes and to make referring to them outside the class easier (e.g. just use gameVars.terrainTextures[0] instead of searching through a list to find the name of the specific one you want).

Drawer: Added a drawer class that will handle drawing things to the screen. Other draw methods will be added later, but currently it just handles drawing the map (and the units on it).

Game1: Simplified interactions with the Level class, refined movement. Units are now restricted to moving one space at a time, and can only move into unoccupied, passable terrain spaces.

CharacterCreator.Designer: Changed textbox "Name" to "UnitName" so things would compile; either VS or SourceTree got confused and put in this.Name instead of this.Name.Name.

General: For testing purposes, I changed the general Grid structure to a 3-piece Terrain/Players/Enemies structure. We will eventually need to consolidate these, i.e. make it so that map files contain *all* GridOccupant data. We need to store this data in the map anyway (to know, for instance, if this terrain is impassable), so it would make sense to store everything together and just have to keep track of one external file for each level. We can talk this out on Monday's meeting if you want a more in-depth explanation. Also, as mentioned a little bit above, terrain index can be removed, as the current Level/Draw setup makes it unnecessary.

P.S. Who am I kidding, none of you read all of that.